sources/requirements.txt
cv2
pandas
tkinter
pyzbar
Pillow
pygame
